updated activity.py
